+++ shared-components/statviz/Components/Visualizations/MovedBoxes/MovedBoxesFilterContainer.tsx
@@ -1,46 +1,15 @@
 import { useMemo } from "react";
-import { Wrap, FormLabel, Box } from "@chakra-ui/react";
-import { zodResolver } from "@hookform/resolvers/zod/src/zod.js";
-import { z } from "zod";
-import SelectField from "../../../../form/SelectField";
 import useTimerange from "../../../hooks/useTimerange";
-import {
-  MovedBoxesData,
-  MovedBoxesResult,
-  TargetDimensionInfo,
-} from "../../../../types/generated/graphql";
+import { MovedBoxesData, MovedBoxesResult } from "../../../../types/generated/graphql";
 import { filterListByInterval } from "../../../../utils/helpers";
 import MovedBoxesCharts from "./MovedBoxesCharts";
-import useListFilter from "../../../hooks/useListFilter";
-
-const singleSelectOptionSchema = z.object({
-  label: z.string(),
-  value: z.string(),
-});
-
-export const MovedBoxesFilterSchema = z.object({
-  locations: singleSelectOptionSchema.array().optional(),
-});
-
-export type IMovedBoxesFilterInput = z.input<typeof MovedBoxesFilterSchema>;
-export type IMovedBoxesFilterOutput = z.output<typeof MovedBoxesFilterSchema>;
+import useValueFilter from "../../../hooks/useValueFilter";
+import { boxesOrItemsFilterValues, defaultBoxesOrItems } from "../../filter/BoxesOrItemsSelect";
 
 export default function MovedBoxesFilterContainer(props: { movedBoxes: MovedBoxesData }) {
   const { interval } = useTimerange();
 
-  const mapTargetToSelectableLocation = (target: TargetDimensionInfo) => ({
-    value: target.id ?? "",
-    label: target.name ?? "",
-  });
-
-  const locations = props.movedBoxes.dimensions?.target?.map(mapTargetToSelectableLocation) ?? [];
-
-  const { control, errors } = useListFilter(
-    "locations",
-    "locations",
-    locations,
-    zodResolver(MovedBoxesFilterSchema),
-  );
+  const { filterValue } = useValueFilter(boxesOrItemsFilterValues, defaultBoxesOrItems, "boi");
 
   const movedBoxesFacts = useMemo(() => {
     try {
@@ -59,24 +28,5 @@ export default function MovedBoxesFilterContainer(props: { movedBoxes: MovedBoxe
     facts: movedBoxesFacts,
     dimensions: props.movedBoxes.dimensions,
   };
-  return (
-    <>
-      <Wrap>
-        <Box width="250px">
-          <FormLabel htmlFor="box-item-select">Categories</FormLabel>
-          <SelectField
-            fieldId="locations"
-            fieldLabel="locations"
-            placeholder="select locations"
-            options={locations}
-            isRequired={false}
-            isMulti
-            errors={errors}
-            control={control}
-          />
-        </Box>
-      </Wrap>
-      <MovedBoxesCharts movedBoxes={filteredMovedBoxesCube} />
-    </>
-  );
+  return <MovedBoxesCharts movedBoxes={filteredMovedBoxesCube} boxesOrItems={filterValue.value} />;
 }
